move updateAllowEmailShare backend logic to core package
