+++ apps/member-profile/app/routes/_profile.profile.emails.tsx
@@ -28,13 +28,9 @@ import {
 } from '../shared/components/profile';
 
 import { Route } from '../shared/constants';
-import { db } from '../shared/core.server';
+import { db, updateAllowEmailShare } from '../shared/core.server';
 import { track } from '../shared/mixpanel.server';
-import {
-  getMember,
-  listEmails,
-  updateAllowEmailShare,
-} from '../shared/queries';
+import { getMember, listEmails } from '../shared/queries';
 import {
   commitSession,
   ensureUserAuthenticated,
