+++ apps/member-profile/app/shared/queries/index.ts
@@ -208,15 +208,3 @@ export async function listEmails(id: string) {
 
   return rows;
 }
-
-export async function updateAllowEmailShare(
-  trx: Transaction<DB>,
-  id: string,
-  allowEmailShareBool: boolean
-) {
-  await trx
-    .updateTable('students')
-    .set({ allowEmailShare: allowEmailShareBool })
-    .where('id', '=', id)
-    .execute();
-}
