+++ packages/core/src/member-profile.server.ts
@@ -34,6 +34,7 @@ export { upsertIcebreakerResponses } from './modules/icebreaker/use-cases/upsert
 export { getAutocompletedCities } from './modules/location/queries/get-autocompleted-cities';
 export { getCityDetails } from './modules/location/queries/get-city-details';
 export { joinMemberDirectory } from './modules/member/use-cases/join-member-directory';
+export { updateAllowEmailShare } from './modules/member/use-cases/update-allow-email-share';
 export { reportError } from './modules/sentry/use-cases/report-error';
 export { countMessagesSent } from './modules/slack/queries/count-messages-sent';
 export { claimSwagPack } from './modules/swag-pack/use-cases/claim-swag-pack';
