+++ packages/core/src/modules/member/use-cases/update-allow-email-share.ts
@@ -1,0 +1,14 @@
+import { Transaction } from 'kysely';
+import { DB } from 'kysely-codegen/dist/db';
+
+export async function updateAllowEmailShare(
+  trx: Transaction<DB>,
+  id: string,
+  allowEmailShareBool: boolean
+) {
+  await trx
+    .updateTable('students')
+    .set({ allowEmailShare: allowEmailShareBool })
+    .where('id', '=', id)
+    .execute();
+}
